Add app installation action
